extract codec funcs to shared dependency-light location
