+++ pkg/solana/logpoller/types/types.go
@@ -4,10 +4,8 @@ import (
 	"context"
 	"crypto/sha256"
 	"database/sql/driver"
-	"encoding/binary"
 	"encoding/json"
 	"fmt"
-	"math"
 	"reflect"
 	"slices"
 	"strings"
@@ -221,25 +219,10 @@ func scanJSON(name string, dest, src interface{}) error {
 // postgres field. Maps, structs, and slices or arrays (of anything but byte) are not supported. For signed
 // or unsigned integer types, strings, or byte arrays, the SQL operators <, =, & > should work in the expected
 // way.
-type IndexedValue []byte
+type IndexedValue = solcommoncodec.IndexedValue
 
-func (v *IndexedValue) FromUint64(u uint64) {
-	*v = make([]byte, 8)
-	binary.BigEndian.PutUint64(*v, u)
-}
-
-func (v *IndexedValue) FromInt64(i int64) {
-	// Golang signed integers are two's complement encoded, so the value will be stored with two's complement encoding.
-	// This also matches the EVM implementation that stores raw topics, EVM ABI encoding also uses two's complement
-	v.FromUint64(uint64(i)) // nolint gosec two's complement encoding
-}
-
-func (v *IndexedValue) FromFloat64(f float64) {
-	if f > 0 {
-		v.FromUint64(math.Float64bits(f) + math.MaxInt64 + 1)
-		return
-	}
-	v.FromUint64(math.MaxInt64 + 1 - math.Float64bits(f))
+func NewIndexedValue(typedVal any) (IndexedValue, error) {
+	return solcommoncodec.NewIndexedValue(typedVal)
 }
 
 type IndexedValues []IndexedValue
@@ -266,52 +249,6 @@ func (v IndexedValues) Value() (driver.Value, error) {
 	}
 
 	return byteArray.Value()
-}
-
-func NewIndexedValue(typedVal any) (iVal IndexedValue, err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = fmt.Errorf("panic recovered: %v while creating indexedValue for %T", r, typedVal)
-		}
-	}()
-	// handle 2 simplest cases first
-	switch t := typedVal.(type) {
-	case []byte:
-		return t, nil
-	case string:
-		return []byte(t), nil
-	}
-
-	// handle numeric types
-	v := reflect.ValueOf(typedVal)
-	if v.CanUint() {
-		iVal.FromUint64(v.Uint())
-		return iVal, nil
-	}
-	if v.CanInt() {
-		iVal.FromInt64(v.Int())
-		return iVal, nil
-	}
-	if v.CanFloat() {
-		iVal.FromFloat64(v.Float())
-		return iVal, nil
-	}
-
-	// any length array is fine as long as the element type is byte
-	if t := v.Type(); t.Kind() == reflect.Array {
-		if t.Elem().Kind() == reflect.Uint8 {
-			if v.CanAddr() {
-				return v.Bytes(), nil
-			}
-			result := make([]byte, v.Len())
-			l := v.Len()
-			for i := 0; i < l; i++ {
-				result[i] = byte(v.Index(i).Uint())
-			}
-			return result, nil
-		}
-	}
-	return nil, fmt.Errorf("can't create indexed value from type %T", typedVal)
 }
 
 type ReplayStatus int
